More Problems
I've added a template to the app, which is fine.  I still don't have it on the wrong user page, but that's a small thing.  What is happening, is that when I add a new secret now, I am being brought back to the login page.  When I then login, I am being told that I have no secrets.  On azure, I am getting a different error now - Cannot read property 'push' of null.  I think it is to do with line 57 of my index.js page, but I'm unsure what to do with it.