No further forward
Changing  if(req.session.userName === undefined || req.session.userName == null) to the original line if(!req.session), fixes all of my issues except getting the app working on azure for the first time.  It also doesn't allow me to logout of the session. I need this line for the app to run on azure, but I'm stumped as to how to fix the other issues.