Starting again
All my secret app project is working perfectly locally, but it's not working properly on azure.  Only the login and wrong login pages are working.  I'm not getting any worthwhile error logs on azure really, so I'm hoping that working with a database rather than an array will solve a lot of issues for me.