Database up and running
This is what I did:  I installed monk, it makes mongo db a little simpler looking in the router.  I followed a tutorial online to create a simple database initially, because it wasn't working for me when I was using mongo with robomongo.  I had created the path and all that jazz in the cmd prompt to store databases in my C drive, but this just wasn't working.  Using the online tutorial finally helped and things seem to be working now.  Took hours, but worth it