Login authentication now working
I retrived all users from the db whose username was the same as what was typed in the form.  I then checked the users password against the password that was typed into the form. There were are a few issues initally, where I mixed up naming conventions, which is silly, but I'm tired :)D