I think things are fixed again
Removing an if statement and deleting all entries from the database helped to fix the problem.  There still seems to be one issue, if a user doesn't already exist, and puts in a username and password, everything 'hangs', nothing happens.  This is to do with the line of code I took out to test for this