Added delete functionality
I was wrong about the issue RE: getting id for deleting from mongo auto generater id. This actually works perfectly. I thought that this was what crashed my app earlier but this was just a coincedince. As mentioned earlier, it was actually due to an if statement checking if results from database is empty. Delete functionally working now.