Edit is now working
Firstly, I was not picking up the right id for editing. I fixed this by adding a new input field which I would send up with the edit form. This means I could grab it from inside my route instead og grabbing it from the query. I think this was not working because this was a post request and not a get request. My secrets then appeared to delete after updating. The reason for this was that I was not including the author when updating the secret so it was not appearing in the list when the new list of updated secrets were being shown from the database after the page had refreshed.